Fix a warning of the unused needsUpdate field.
When importing a fresh SpectatorView pull, you'll see this warning in the Unity console because this field is only used when STATESYNC_TEXTMESHPRO is specified. Of course when you're opening things for the first time (i.e. ingesting SV for the first time) it's possible to have made this step (because adding that define is a step much later)

This just puts it guarded by the define that it's actually used in.